+++ run_ifcb_est.py
@@ -2,7 +2,6 @@
 from gnss_config import GNSSconfig
 from gnss_time import GNSStime, hms2sod
 import gnss_tools as gt
-# import gnss_files as gf
 import gnss_run as gr
 from constants import read_site_list
 import os
@@ -15,23 +14,26 @@ logging.basicConfig(level=logging.DEBUG,
                     format='%(asctime)s - %(filename)s[line:%(lineno)d] - %(levelname)s: %(message)s')
 
 # ------ Get args ----------------
-parser = argparse.ArgumentParser(description='Run Precise Clock Estimating')
+parser = argparse.ArgumentParser(description='Run IFCB Estimation')
+# Time argument
 parser.add_argument('-n', dest='num', type=int, default=1, help='number of process days')
 parser.add_argument('-l', dest='len', type=int, default=24, help='process time length (hours)')
-parser.add_argument('-i', dest='intv', type=int, default=300, help='process interval (seconds)')
-parser.add_argument('-c', dest='obs_comb', default='IF', choices={'UC', 'IF'}, help='Observation combination')
-parser.add_argument('-est', dest='est', default='LSQ', choices={'EPO', 'LSQ'}, help='Estimator: LSQ or EPO')
+parser.add_argument('-i', dest='intv', type=int, default=30, help='process interval (seconds)')
+parser.add_argument('-t', dest='hms', nargs='+', help='begin date: hh mm ss')
+parser.add_argument('-sod', dest='sod', help='begin date: seconds of day')
+# Estimation argument
 parser.add_argument('-sys', dest='sys', default='G', help='used GNSS observations, e.g. G/GC/GREC')
-parser.add_argument('-freq', dest='freq', type=int, default=3, help='used GNSS frequencies')
-parser.add_argument('-cen', dest='cen', default='com', choices={'igs', 'cod', 'com', 'wum', 'gbm', 'grm', 'sgg'},
+# File argument
+parser.add_argument('-cen', dest='cen', default='com', choices={'igs', 'cod', 'com', 'wum', 'gbm', 'grm', 'sgg', 'grt'},
                     help='GNSS precise orbits and clocks')
 parser.add_argument('-bia', dest='bia', default='cas', choices={'cod', 'cas', 'whu', 'sgg'},
                     help='bias files')
+parser.add_argument('-cf', dest='cf', default='cf_upd.ini', help='config file')
+parser.add_argument('-kp', dest='keep_dir', action='store_true', help='Keep the existing work dir')
+# Required argument
 parser.add_argument('-s', dest='f_list', required=True, help='site_list file')
 parser.add_argument('-y', dest='year', type=int, required=True, help='begin date: year')
 parser.add_argument('-d', dest='doy', type=int, required=True, help='begin date: day of year')
-parser.add_argument('-t', dest='hms', nargs='+', help='begin date: hh mm ss')
-parser.add_argument('-sod', dest='sod', help='begin date: seconds of day')
 args = parser.parse_args()
 
 # ------ Path information --------
@@ -53,29 +55,23 @@ else:
 # ------ Init config file --------
 sta_list = read_site_list(args.f_list)
 sta_list.sort()
-f_config_tmp = 'pce_config.ini'
-config = GNSSconfig(f_config_tmp)
+if not os.path.isfile(args.cf):
+    raise SystemExit("Cannot get config file >_<")
+config = GNSSconfig(args.cf)
 config.update_pathinfo(sys_data, gns_data, upd_data)
-config.update_gnssinfo(args.sys, args.freq, args.obs_comb, args.est)
+config.update_gnssinfo(args.sys, 3, "IF", "LSQ")
 if sta_list:
     config.update_stalist(sta_list)
 else:
     raise SystemExit("No site to process")
-if args.freq > 2:
-    args.bia = "CAS"
 config.update_prodinfo(args.cen, args.bia)
 
 # ------ Start PPP process -------
-proj_dir = os.path.join(base_dir, 'PCE')
+proj_dir = os.path.join(base_dir, 'UPD')
 if args.sod:
     sod = args.sod
 elif args.hms:
-    if len(args.hms) > 2:
-        sod = hms2sod(args.hms[0], args.hms[1], args.hms[2])
-    elif len(args.hms) > 1:
-        sod = hms2sod(args.hms[0], args.hms[1])
-    else:
-        sod = hms2sod(args.hms[0])
+    sod = hms2sod(args.hms[0:3])
 else:
     sod = hms2sod(0)
 count = args.num
@@ -87,21 +83,19 @@ while count > 0:
     t_beg = t_beg0
     t_end = t_beg.time_increase(seslen-args.intv)
     config.update_timeinfo(t_beg, t_end, args.intv)
-    config.update_process(crd_constr='FIX')
-    logging.info(f"\n===> Run PCE for {t_beg.year}-{t_beg.doy:0>3d}\n")
-    workdir = os.path.join(proj_dir, str(t_beg.year), f"{t_beg.doy:0>3d}_{args.sys}_{args.freq}_{args.obs_comb}")
+    logging.info(f"\n===> Run IFCB for {t_beg.year}-{t_beg.doy:0>3d}\n")
+    workdir = os.path.join(proj_dir, str(t_beg.year), f"{t_beg.doy:0>3d}_{args.sys}_ifcb")
     if not os.path.isdir(workdir):
         os.makedirs(workdir)
     else:
-        shutil.rmtree(workdir)
-        os.makedirs(workdir)
+        if not args.keep_dir:
+            shutil.rmtree(workdir)
+            os.makedirs(workdir)
     os.chdir(workdir)
-    gt.mkdir(['log_tb', 'clkdif', 'tmp'])
     logging.info(f"work directory is {workdir}")
 
     # ---------- Basic check ---------
-    config.copy_sys_data()
-    if config.basic_check(['estimator'], ['rinexo', 'rinexn', 'sp3', 'biabern']):
+    if config.basic_check(['estimator'], ['rinexo', 'biabern']):
         logging.info("Basic check complete ^_^")
     else:
         logging.critical("Basic check failed! skip to next day")
@@ -112,11 +106,10 @@ while count > 0:
     f_config = os.path.join(workdir, 'config.ini')
     config.write_config(f_config)  # config file is only for check
     logging.info(f"config is {f_config}")
+
     # Run turboedit
-    config.update_process(intv=30)
     nthread = min(len(config.all_receiver().split()), 10)
     gr.run_great(grt_bin, 'great_turboedit', config, nthread=nthread, out=os.path.join("tmp", "turboedit"))
-    config.update_process(intv=args.intv)
     if config.basic_check(files=['ambflag']):
         logging.info("Ambflag is ok ^_^")
     else:
@@ -124,9 +117,10 @@ while count > 0:
         t_beg = t_beg.time_increase(86400)
         count -= 1
         continue
-    # Run Precise Clock Estimation
-    gr.run_great(grt_bin, 'great_pcelsq', config, mode='PCE_EST', out=os.path.join("tmp", "pcelsq"))
-    gr.run_great(grt_bin, 'great_clkdif', config)
+    # get ifcb for GPS
+    gr.run_great(grt_bin, 'great_updlsq', config, mode='ifcb', out="ifcb")
+
+    # gt.copy_result_files_to_path(config, ["ifcb"], os.path.join(upd_data, f"{t_beg.year}"))
 
     # next day
     logging.info(f"Complete {t_beg.year}-{t_beg.doy:0>3d} ^_^\n")
