+++ run_ppp.py
@@ -16,32 +16,38 @@ logging.basicConfig(level=logging.DEBUG,
 
 # ------ Get args ----------------
 parser = argparse.ArgumentParser(description='Perform Precise Point Positioning')
+# Time argument
 parser.add_argument('-n', dest='num', type=int, default=1, help='number of process days')
 parser.add_argument('-l', dest='len', type=int, default=24, help='process time length (hours)')
 parser.add_argument('-i', dest='intv', type=int, default=30, help='process interval (seconds)')
-parser.add_argument('-c', dest='obs_comb', default='IF', choices={'UC', 'IF'}, help='Observation combination')
-parser.add_argument('-est', dest='est', default='EPO', choices={'EPO', 'LSQ'}, help='Estimator: LSQ or EPO')
+parser.add_argument('-t', dest='hms', nargs='+', help='begin date: hh mm ss')
+parser.add_argument('-sod', dest='sod', help='begin date: seconds of day')
+# Estimation argument
+parser.add_argument('-c', dest='obs_comb', default='IF', choices={'UC', 'IF'}, help='observation combination')
+parser.add_argument('-est', dest='est', default='EPO', choices={'EPO', 'LSQ'}, help='estimator: LSQ or EPO')
 parser.add_argument('-sys', dest='sys', default='G', help='used GNSS observations, e.g. G/GC/GREC')
-parser.add_argument('-freq', dest='freq', type=int, default=3, help='used GNSS frequencies')
-parser.add_argument('-cen', dest='cen', default='com', choices={'igs', 'cod', 'com', 'wum', 'gbm', 'grm', 'sgg'},
+parser.add_argument('-freq', dest='freq', type=int, default=2, help='used GNSS frequencies')
+# File argument
+parser.add_argument('-cen', dest='cen', default='com', choices={'igs', 'cod', 'com', 'wum', 'gbm', 'grm', 'sgg', 'grt'},
                     help='GNSS precise orbits and clocks')
 parser.add_argument('-bia', dest='bia', default='cas', choices={'cod', 'cas', 'whu', 'sgg'},
                     help='bias files')
+parser.add_argument('-cf', dest='cf', default='cf_ppp.ini', help='config file')
+parser.add_argument('-kp', dest='keep_dir', action='store_true', help='Keep the existing work dir')
+# Required argument
 parser.add_argument('-s', dest='f_list', required=True, help='site_list file')
 parser.add_argument('-y', dest='year', type=int, required=True, help='begin date: year')
 parser.add_argument('-d', dest='doy', type=int, required=True, help='begin date: day of year')
-parser.add_argument('-t', dest='hms', nargs='+', help='begin date: hh mm ss')
-parser.add_argument('-sod', dest='sod', help='begin date: seconds of day')
 args = parser.parse_args()
 
 # ------ Path information --------
 if platform.system() == 'Windows':
-    grt_dir = r"C:\Users\jiaqi\GNSS_Software\branches"
-    grt_bin = os.path.join(grt_dir, 'merge_navpod_merge_ppp', 'build', 'Bin', 'RelWithDebInfo')
-    sys_data = r"C:\Users\jiaqi\GNSS_Project\sys_data"
-    gns_data = r"C:\Users\jiaqi\GNSS_Project\gns_data"
-    upd_data = r"C:\Users\jiaqi\GNSS_Project\gns_data\upd"
-    base_dir = r"C:\Users\jiaqi\GNSS_Project"
+    grt_dir = r"D:\GNSS_Software\GREAT"
+    grt_bin = os.path.join(grt_dir, 'build', 'Bin', 'RelWithDebInfo')
+    sys_data = r"D:\GNSS_Project\sys_data"
+    gns_data = r"D:\GNSS_Project\gns_data"
+    upd_data = r"D:\GNSS_Project\gns_data\upd"
+    base_dir = r"D:\GNSS_Project"
 else:
     grt_dir = "/home/jqwu/softwares/GREAT/branches"
     grt_bin = os.path.join(grt_dir, 'merge_navpod_merge_ppp', 'build', 'Bin')
@@ -53,14 +59,14 @@ else:
 # ------ Init config file --------
 sta_list = read_site_list(args.f_list)
 sta_list.sort()
-f_config_tmp = 'ppp_config.ini'
-config = GNSSconfig(f_config_tmp)
+if not sta_list:
+    raise SystemExit("No site to process")
+if not os.path.isfile(args.cf):
+    raise SystemExit("Cannot get config file >_<")
+config = GNSSconfig(args.cf)
 config.update_pathinfo(sys_data, gns_data, upd_data)
 config.update_gnssinfo(args.sys, args.freq, args.obs_comb, args.est)
-if sta_list:
-    config.update_stalist(sta_list)
-else:
-    raise SystemExit("No site to process")
+config.update_stalist(sta_list)
 config.update_prodinfo(args.cen, args.bia)
 
 # ------ Start PPP process -------
@@ -91,8 +97,9 @@ while count > 0:
     if not os.path.isdir(workdir):
         os.makedirs(workdir)
     else:
-        shutil.rmtree(workdir)
-        os.makedirs(workdir)
+        if not args.keep_dir:
+            shutil.rmtree(workdir)
+            os.makedirs(workdir)
     os.chdir(workdir)
     gt.mkdir(['log_tb', 'enu', 'flt', 'ppp', 'ambupd', 'res', 'tmp'])
     logging.info(f"work directory is {workdir}")
@@ -136,17 +143,17 @@ while count > 0:
     gr.run_great(grt_bin, 'great_ppplsq', config, mode='PPP_EST', newxml=True, nthread=nthread, fix_mode="SEARCH",
                  out=os.path.join('tmp', 'ppplsq'))
 
-    # config.update_gnssinfo(obs_model='IF', freq=3)
-    # gr.run_great(grt_bin, 'great_ppplsq', config, mode='PPP_EST', newxml=True, nthread=nthread, fix_mode="NO",
-    #              out=os.path.join('tmp', 'ppplsq'))
-    # gr.run_great(grt_bin, 'great_ppplsq', config, mode='PPP_EST', newxml=True, nthread=nthread, fix_mode="SEARCH",
-    #              out=os.path.join('tmp', 'ppplsq'))
-    #
-    # config.update_gnssinfo(obs_model='IF', freq=2)
-    # gr.run_great(grt_bin, 'great_ppplsq', config, mode='PPP_EST', newxml=True, nthread=nthread, fix_mode="NO",
-    #              out=os.path.join('tmp', 'ppplsq'))
-    # gr.run_great(grt_bin, 'great_ppplsq', config, mode='PPP_EST', newxml=True, nthread=nthread, fix_mode="SEARCH",
-    #              out=os.path.join('tmp', 'ppplsq'))
+    config.update_gnssinfo(obs_model='IF', freq=3)
+    gr.run_great(grt_bin, 'great_ppplsq', config, mode='PPP_EST', newxml=True, nthread=nthread, fix_mode="NO",
+                 out=os.path.join('tmp', 'ppplsq'))
+    gr.run_great(grt_bin, 'great_ppplsq', config, mode='PPP_EST', newxml=True, nthread=nthread, fix_mode="SEARCH",
+                 out=os.path.join('tmp', 'ppplsq'))
+
+    config.update_gnssinfo(obs_model='IF', freq=2)
+    gr.run_great(grt_bin, 'great_ppplsq', config, mode='PPP_EST', newxml=True, nthread=nthread, fix_mode="NO",
+                 out=os.path.join('tmp', 'ppplsq'))
+    gr.run_great(grt_bin, 'great_ppplsq', config, mode='PPP_EST', newxml=True, nthread=nthread, fix_mode="SEARCH",
+                 out=os.path.join('tmp', 'ppplsq'))
 
     # next day
     logging.info(f"Complete {t_beg.year}-{t_beg.doy:0>3d} ^_^\n")
